+++ src/worker.rs
@@ -11681,6 +11681,7 @@ pub(super) fn sample_account() -> AccountConfig {
         empty_trash_days: 0,
         pgp_key: None,
         in_unified: true,
+        folder_sort: None,
         sign_by_default: false,
         push: None,
         name: String::new(),